Edited project1.py, did average wait time wrong
Im dumb and did average wait time wrong, I did it before as the average serve time between the max/min. Now its the average time it took to each customer that HAS been served. Doesn't apply to those who are still waiting. Look it over im not sure if this is the best solution